chore: move golang codebase to `pkg` directory
