+++ pkg/internal/constants/meta/script.js
