+++ pkg/internal/utils/utils_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/vertexcover-io/locatr/golang/types"
+	"github.com/vertexcover-io/locatr/pkg/types"
 )
 
 func TestParseElementSpec(t *testing.T) {
